Only try to delete duplicates if there is any

+++ communicator.py
@@ -206,13 +206,14 @@ def process_as_feedback(feedback):
         conn.commit()
         df = pd.read_sql("SELECT * FROM attempted_matches", conn)
         delete_records = df[df.duplicated(subset=["project_id", "cert_id"], keep='last')]
-        for record in delete_records:
-            conn.cursor().execute(f"""
-                DELETE FROM attempted_matches
-                WHERE idx = %s
-            """, [record.idx]
-            )
-        conn.commit()
+        if len(delete_records):
+            for record in delete_records:
+                conn.cursor().execute(f"""
+                    DELETE FROM attempted_matches
+                    WHERE idx = %s
+                """, [record.idx]
+                )
+            conn.commit()
         logger.info(
             f"cert_id`{cert_id}` from {source} was a "
             f"{'successful match' if response == 1 else 'mis-match'} for job "
